Move taxonomy lists into scroll areas
