+++ naturtag/app/taxon_controller.py
@@ -1,26 +1,16 @@
 from logging import getLogger
 from time import time
-from typing import Callable, Iterator
+from typing import Callable, Iterable, Iterator
 
 from pyinaturalist import Taxon
 from PySide6.QtCore import Qt, Signal
 from PySide6.QtGui import QFont
-from PySide6.QtWidgets import (
-    QGroupBox,
-    QHBoxLayout,
-    QLabel,
-    QLayout,
-    QLineEdit,
-    QListWidget,
-    QVBoxLayout,
-    QWidget,
-)
+from PySide6.QtWidgets import QGroupBox, QLabel, QLineEdit, QListWidget, QScrollArea, QWidget
 
-from naturtag.app.image_window import PixmapLabel
+from naturtag.app.images import PixmapLabel
+from naturtag.app.layouts import HorizontalLayout, VerticalLayout
 from naturtag.metadata.inat_metadata import INAT_CLIENT
 from naturtag.settings import Settings
-
-# from naturtag.thumbnails import get_thumbnail
 
 logger = getLogger(__name__)
 
@@ -33,35 +23,22 @@ H2_FONT.setWeight(QFont.Bold)
 H2_FONT.setPointSize(16)
 
 
-ancestors = [
-    Taxon(name='Animalia', id=1, rank='kingdom', preferred_common_name='Animals'),
-    Taxon(name='Arthropoda', id=1, rank='phylum', preferred_common_name='Arthropods'),
-    Taxon(name='Hexapoda', id=1, rank='subphylum', preferred_common_name='Hexapods'),
-    Taxon(name='Insecta', id=1, rank='class', preferred_common_name='Insects'),
-    Taxon(name='Pterygota', id=1, rank='subclass', preferred_common_name='Winged insects'),
-]
-children = [
-    Taxon(name='Anisoptera', id=12345, rank='suborder', preferred_common_name='Dragonflies'),
-    Taxon(name='Zygoptera', id=12345, rank='suborder', preferred_common_name='Damselflies'),
-]
-
-
 class TaxonController(QWidget):
     """Controller for searching taxa"""
 
     def __init__(self, settings: Settings, info_callback: Callable):
         super().__init__()
         self.settings = settings
-        root_layout = QHBoxLayout()
+        root_layout = HorizontalLayout()
         self.setLayout(root_layout)
         self.selected_taxon: Taxon = None
         self.info = info_callback
 
-        input_layout = QVBoxLayout()
+        input_layout = VerticalLayout()
         root_layout.addLayout(input_layout)
 
         # Taxon name autocomplete
-        autocomplete_layout = QVBoxLayout()
+        autocomplete_layout = VerticalLayout()
         autocomplete_layout.setAlignment(Qt.AlignTop)
         group_box = QGroupBox('Search')
         group_box.setFixedWidth(400)
@@ -80,7 +57,7 @@ class TaxonController(QWidget):
         autocomplete_layout.addWidget(autocomplete_results)
 
         # Iconic taxa (category) inputs
-        categories_layout = QVBoxLayout()
+        categories_layout = VerticalLayout()
         group_box = QGroupBox('Categories')
         group_box.setFixedWidth(400)
         group_box.setFont(H2_FONT)
@@ -92,7 +69,7 @@ class TaxonController(QWidget):
         categories_layout.addWidget(QLabel('3'))
 
         # Rank inputs
-        rank_layout = QVBoxLayout()
+        rank_layout = VerticalLayout()
         group_box = QGroupBox('Rank')
         group_box.setFixedWidth(400)
         group_box.setFont(H2_FONT)
@@ -106,7 +83,7 @@ class TaxonController(QWidget):
         # -----------------
 
         # Selected taxon
-        results_layout = QVBoxLayout()
+        results_layout = VerticalLayout()
         root_layout.addLayout(results_layout)
 
         self.selected_taxon_title = QLabel('Selected Taxon')
@@ -115,7 +92,6 @@ class TaxonController(QWidget):
 
         self.taxon_info = TaxonInfoSection()
         results_layout.addLayout(self.taxon_info)
-
         self.taxonomy = TaxonomySection()
         results_layout.addLayout(self.taxonomy)
 
@@ -124,14 +100,15 @@ class TaxonController(QWidget):
     def select_taxon(self, taxon_id: int = None, taxon: Taxon = None):
         """Update taxon info display"""
         # Don't need to do anything if this taxon is already selected
-        id = taxon_id or taxon.id
-        if self.selected_taxon is not None and id == self.selected_taxon.id:
+        id = taxon_id or getattr(taxon, 'id', None)
+        if self.selected_taxon and self.selected_taxon.id == id:
             return
 
         logger.info(f'Selecting taxon {id}')
         start = time()
         if taxon_id and not taxon:
             taxon = INAT_CLIENT.taxa.from_id(taxon_id).one()
+        assert taxon is not None
         self.selected_taxon = taxon
 
         common_name = f' ({taxon.preferred_common_name}) ' if taxon.preferred_common_name else ''
@@ -141,12 +118,12 @@ class TaxonController(QWidget):
         self.taxonomy.load(taxon)
 
         # Clicking on a taxon card will select it
-        for card in self.taxonomy.taxon_cards:
+        for card in self.taxonomy.taxa:
             card.clicked.connect(self.select_taxon)
-        logger.info(f'Loaded taxon {taxon.id} {time() - start:.2f}s')
+        logger.info(f'Loaded taxon {taxon.id} in {time() - start:.2f}s')
 
 
-class TaxonInfoSection(QHBoxLayout):
+class TaxonInfoSection(HorizontalLayout):
     """Section to display selected taxon photo and basic info"""
 
     def __init__(self):
@@ -159,12 +136,15 @@ class TaxonInfoSection(QHBoxLayout):
         self.image.setMaximumWidth(400)
         self.addWidget(self.image)
 
+        self.icon_layout = HorizontalLayout()
+        self.icon_layout.setAlignment(Qt.AlignTop)
         self.icon = PixmapLabel()
         self.icon.setFixedSize(75, 75)
+        self.icon_layout.addWidget(self.icon)
+        self.addLayout(self.icon_layout)
 
-        self.details = QVBoxLayout()
+        self.details = VerticalLayout()
         self.details.setAlignment(Qt.AlignTop)
-        self.details.addWidget(self.icon)
         self.addLayout(self.details)
 
     def load(self, taxon: Taxon):
@@ -173,34 +153,29 @@ class TaxonInfoSection(QHBoxLayout):
         self.icon.setPixmap(url=taxon.icon_url)
 
         # Other attributes
+        self.details.clear()
         self.details.addWidget(QLabel(f'ID: {taxon.id}'))
         self.details.addWidget(QLabel(f'Rank: {taxon.rank}'))
         self.details.addWidget(QLabel(f'Observations: {taxon.observations_count}'))
         self.details.addWidget(QLabel(f'Child species: {taxon.complete_species_count}'))
 
 
-class TaxonomySection(QHBoxLayout):
+class TaxonomySection(HorizontalLayout):
     """Section to display ancestors and children of selected taxon"""
 
     def __init__(self):
         super().__init__()
 
-        # Ancestors
-        self.ancestors_layout = QVBoxLayout()
-        self.ancestors_layout.setAlignment(Qt.AlignTop)
         self.ancestors_group = QGroupBox('Ancestors')
         self.ancestors_group.setFixedWidth(400)
         self.ancestors_group.setFont(H2_FONT)
-        self.ancestors_group.setLayout(self.ancestors_layout)
+        self.ancestors_layout = TaxonList(self.ancestors_group)
         self.addWidget(self.ancestors_group)
 
-        # Children
-        self.children_layout = QVBoxLayout()
-        self.children_layout.setAlignment(Qt.AlignTop)
         self.children_group = QGroupBox('Children')
         self.children_group.setFixedWidth(400)
         self.children_group.setFont(H2_FONT)
-        self.children_group.setLayout(self.children_layout)
+        self.children_layout = TaxonList(self.children_group)
         self.addWidget(self.children_group)
 
     def load(self, taxon: Taxon):
@@ -210,26 +185,51 @@ class TaxonomySection(QHBoxLayout):
         def get_label(text: str, items: list) -> str:
             return text + (f' ({len(items)})' if items else '')
 
-        # Load ancestors
         self.ancestors_group.setTitle(get_label('Ancestors', taxon.ancestors))
-        clear_layout(self.ancestors_layout)
-        for t in taxon.ancestors:
-            self.ancestors_layout.addWidget(TaxonInfoCard(taxon=t))
-
-        # Load children
+        self.ancestors_layout.set_taxa(taxon.ancestors)
         self.children_group.setTitle(get_label('Children', taxon.children))
-        clear_layout(self.children_layout)
-        for t in taxon.children:
-            self.children_layout.addWidget(TaxonInfoCard(taxon=t))
+        self.children_layout.set_taxa(taxon.children)
 
     @property
-    def taxon_cards(self) -> Iterator['TaxonInfoCard']:
-        for item in self.ancestors_group.children():
+    def taxa(self) -> Iterator['TaxonInfoCard']:
+        yield from self.ancestors_layout.taxa
+        yield from self.children_layout.taxa
+
+
+class TaxonList(VerticalLayout):
+    """A scrollable list of TaxonInfoCards"""
+
+    def __init__(self, parent: QWidget = None):
+        super().__init__(parent)
+
+        self.scroll_panel = QWidget()
+        self.scroll_layout = VerticalLayout(self.scroll_panel)
+        self.scroll_layout.setAlignment(Qt.AlignTop)
+        self.scroll_layout.setContentsMargins(0, 0, 0, 0)
+        self.addLayout(self.scroll_layout)
+
+        scroll_area = QScrollArea()
+        scroll_area.setWidgetResizable(True)
+        scroll_area.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
+        scroll_area.setWidget(self.scroll_panel)
+        self.addWidget(scroll_area)
+
+    @property
+    def taxa(self) -> Iterator['TaxonInfoCard']:
+        for item in self.scroll_panel.children():
             if isinstance(item, TaxonInfoCard):
                 yield item
-        for item in self.children_group.children():
-            if isinstance(item, TaxonInfoCard):
-                yield item
+
+    def add_taxon(self, taxon: Taxon):
+        self.scroll_layout.addWidget(TaxonInfoCard(taxon=taxon))
+
+    def clear(self):
+        self.scroll_layout.clear()
+
+    def set_taxa(self, taxa: Iterable[Taxon]):
+        self.clear()
+        for taxon in taxa:
+            self.add_taxon(taxon)
 
 
 class TaxonInfoCard(QWidget):
@@ -237,7 +237,7 @@ class TaxonInfoCard(QWidget):
 
     def __init__(self, taxon: Taxon):
         super().__init__()
-        card_layout = QHBoxLayout()
+        card_layout = HorizontalLayout()
         self.setLayout(card_layout)
         self.taxon_id = taxon.id
         print('Taxon ID:', self.taxon_id)
@@ -250,7 +250,7 @@ class TaxonInfoCard(QWidget):
         # Details
         title = QLabel(taxon.name)
         title.setFont(H2_FONT)
-        details_layout = QVBoxLayout()
+        details_layout = VerticalLayout()
         card_layout.addLayout(details_layout)
         details_layout.addWidget(title)
         details_layout.addWidget(QLabel(taxon.rank))
@@ -262,13 +262,3 @@ class TaxonInfoCard(QWidget):
     def mouseReleaseEvent(self, event):
         if event.button() == Qt.LeftButton:
             self.clicked.emit(self.taxon_id)
-
-
-def clear_layout(layout: QLayout):
-    """Why is this not built-in???"""
-    if not layout:
-        return
-    for i in reversed(range(layout.count())):
-        child = layout.takeAt(i)
-        if child.widget():
-            child.widget().deleteLater()
